+++ src/index.ts
@@ -1,4 +1,6 @@
 export { Breadline } from "./breadline";
 export { BreadlineEvent } from "./enum";
 export type { BreadlineOptions, TaskAddOptions } from "./interfaces";
+export { BinaryMaxHeap } from "./queues/binary-max-heap";
+export { PriorityQueue } from "./queues/priority-queue";
 export * from "./types";